Breakup ranger: Warn instead exception for delete Persona roleId not found

+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v2/preprocessor/accesscontrol/PersonaPreProcessor.java
@@ -38,6 +38,7 @@ import org.apache.atlas.type.AtlasType;
 import org.apache.atlas.type.AtlasTypeRegistry;
 import org.apache.atlas.utils.AtlasPerfMetrics;
 import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.lang.StringUtils;
 import org.keycloak.admin.client.resource.GroupResource;
 import org.keycloak.admin.client.resource.GroupsResource;
 import org.keycloak.admin.client.resource.RoleByIdResource;
@@ -313,10 +314,12 @@ public class PersonaPreProcessor implements PreProcessor {
     }
 
     private void removeKeycloakRole(String roleId) {
-        RoleByIdResource rolesResource = KeycloakClient.getKeycloakClient().getRealm().rolesById();
-        //RoleRepresentation roleRepresentation = rolesResource.getRole(roleId);
+        if (StringUtils.isNotEmpty(roleId)) {
+            RoleByIdResource rolesResource = KeycloakClient.getKeycloakClient().getRealm().rolesById();
+            //RoleRepresentation roleRepresentation = rolesResource.getRole(roleId);
 
-        rolesResource.deleteRole(roleId);
+            rolesResource.deleteRole(roleId);
+        }
     }
 
     private AtlasEntity.AtlasEntityWithExtInfo getAccessControlEntity(AtlasEntity entity, EntityMutations.EntityOperation op) throws AtlasBaseException {
